Re-factor Reference-Mapper, to be less link-specific. DDFSAL-189
We have a mapper base, that is used for link-reference fields.
Some of this logic will also need to be used for entity-reference
fields, and instead of just copying the code, we'll re-factor the base
to be more generalized instead.
