+++ web/modules/custom/bnf/src/Plugin/bnf_mapper/FieldBannerLinkMapper.php
@@ -6,6 +6,7 @@ namespace Drupal\bnf\Plugin\bnf_mapper;
 
 use Drupal\bnf\Attribute\BnfMapper;
 use Drupal\bnf\GraphQL\Operations\GetNode\Node\Paragraphs\BannerLink\Link;
+use Spawnia\Sailor\ObjectLike;
 
 /**
  * Mapping Link data.
@@ -13,6 +14,17 @@ use Drupal\bnf\GraphQL\Operations\GetNode\Node\Paragraphs\BannerLink\Link;
 #[BnfMapper(
   id: Link::class,
   )]
-class FieldBannerLinkMapper extends BnfMapperImportedLinkFieldPluginBase {
+class FieldBannerLinkMapper extends BnfMapperImportReferencePluginBase {
+
+  /**
+   * {@inheritDoc}
+   */
+  public function map(ObjectLike $object): mixed {
+    if (!$object instanceof Link) {
+      throw new \RuntimeException('Wrong class handed to mapper');
+    }
+
+    return $this->mapLink($object);
+  }
 
 }
